Refactor Claim Update workflow tests into modular test classes and add support for cancel-entire-request validation
